+++ icd-db/src/main/scala/csw/services/icd/db/parser/EventModelBsonParser.scala
@@ -2,7 +2,7 @@ package csw.services.icd.db.parser
 
 import csw.services.icd.html.HtmlMarkup
 import icd.web.shared.IcdModels.EventModel
-import reactivemongo.bson.BSONDocument
+import reactivemongo.bson.{BSONDocument, BSONNumberLike}
 
 /**
  * See resources/event-schema.conf
@@ -14,11 +14,9 @@ object EventModelBsonParser {
       name = doc.getAs[String]("name").get,
       description = doc.getAs[String]("description").map(HtmlMarkup.gfmToHtml).getOrElse(""),
       requirements = doc.getAs[Array[String]]("requirements").map(_.toList).getOrElse(Nil),
-      minRate = safeNumGet("minRate", doc),
-      maxRate = safeNumGet("maxRate", doc),
+      maybeMaxRate = doc.getAs[BSONNumberLike]("maxRate").map(_.toDouble),
       archive = doc.getAs[Boolean]("archive").getOrElse(false),
       archiveDuration = doc.getAs[String]("archiveDuration").getOrElse(""),
-      archiveRate = safeNumGet("archiveRate", doc),
       attributesList =
         for (subDoc <- doc.getAs[Array[BSONDocument]]("attributes").map(_.toList).getOrElse(Nil))
           yield AttributeModelBsonParser(subDoc),